Move indexer initialization to environment variables
