+++ .testnet.localhost.env
@@ -16,3 +16,6 @@ ALGORAND_EXPLORER=https://api.algoexplorer.io
 ALGORAND_ALGOD_SERVER=http://ec2-3-18-80-65.us-east-2.compute.amazonaws.com
 ALGORAND_ALGOD_PORT=8080
 ALGORAND_TOKEN=11e4dcfb445a8c7e8380848747f18afcd5d84ccb395e003e5b72127ca5e9a259
+
+ALGORAND_INDEXER_SERVER=https://testnet-algorand.api.purestake.io/idx2
+ALGORAND_INDEXER_TOKEN=VELyABA1dGqGbAVktbew4oACvp0c0298gMgYtYIb
